# backoffice-api/api.yaml
openapi: 3.0.3
info:
  title: Shop API
  description: API for Shop
  version: 0.0.3
servers:
  - url: http://localhost:18080/service/api
    description: Generated server url
paths:
...
  /v1/price/update:
    put:
      tags:
        - price
      operationId: updatePrice
      requestBody:
        $ref: '#/components/requestBodies/PriceDto'
      responses:
        200:
          $ref: '#/components/responses/PriceResponse'
        400:
          $ref: '#/components/responses/RequestError'
        401:
          $ref: '#/components/responses/RequestError'
        403:
          $ref: '#/components/responses/RequestError'
        404:
          $ref: '#/components/responses/RequestError'
        500:
          $ref: '#/components/responses/RequestError'
...
  /v1/price/{id}:
    get:
      tags:
        - price
      operationId: getPriceById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          $ref: '#/components/responses/PriceResponse'
        400:
          $ref: '#/components/responses/RequestError'
        401:
          $ref: '#/components/responses/RequestError'
        403:
          $ref: '#/components/responses/RequestError'
        404:
          $ref: '#/components/responses/RequestError'
        500:
          $ref: '#/components/responses/RequestError'
  /v1/price/product:
    get:
      tags:
        - price
      summary: find all prices by product
      operationId: getPricesByProduct
      requestBody:
        $ref: '#/components/requestBodies/ProductDto'
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        200:
          $ref: '#/components/responses/PricesResponse'
        400:
          $ref: '#/components/responses/RequestError'
        401:
          $ref: '#/components/responses/RequestError'
        403:
          $ref: '#/components/responses/RequestError'
        404:
          $ref: '#/components/responses/RequestError'
        500:
          $ref: '#/components/responses/RequestError'
  /v1/price/product/{productId}:
    get:
      tags:
        - price
      summary: find all prices by product
      operationId: getPricesByProductId
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        200:
          $ref: '#/components/responses/PricesResponse'
        400:
          $ref: '#/components/responses/RequestError'
        401:
          $ref: '#/components/responses/RequestError'
        403:
          $ref: '#/components/responses/RequestError'
        404:
          $ref: '#/components/responses/RequestError'
        500:
          $ref: '#/components/responses/RequestError'
  /v1/price/list:
    get:
      tags:
        - price
      operationId: getPrices
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        200:
          $ref: '#/components/responses/PricesResponse'
        400:
          $ref: '#/components/responses/RequestError'
        401:
          $ref: '#/components/responses/RequestError'
        403:
          $ref: '#/components/responses/RequestError'
        404:
          $ref: '#/components/responses/RequestError'
        500:
          $ref: '#/components/responses/RequestError'
  /v1/price/currency/{currency}:
    get:
      tags:
        - price
      summary: find all prices by currency
      operationId: getPricesByCurrency
      parameters:
        - name: currency
          in: path
          required: true
          schema:
            type: string
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        200:
          $ref: '#/components/responses/PricesResponse'
        400:
          $ref: '#/components/responses/RequestError'
        401:
          $ref: '#/components/responses/RequestError'
        403:
          $ref: '#/components/responses/RequestError'
        404:
          $ref: '#/components/responses/RequestError'
        500:
          $ref: '#/components/responses/RequestError'
  /v1/price/between:
    get:
      tags:
        - price
      summary: find all prices between
      operationId: getPricesBetween
      requestBody:
        $ref: '#/components/requestBodies/BetweenPrice'
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        200:
          $ref: '#/components/responses/PricesResponse'
        400:
          $ref: '#/components/responses/RequestError'
        401:
          $ref: '#/components/responses/RequestError'
        403:
          $ref: '#/components/responses/RequestError'
        404:
          $ref: '#/components/responses/RequestError'
        500:
          $ref: '#/components/responses/RequestError'
  /v1/price/between/{currency}/{from}/{to}:
    get:
      tags:
        - price
      summary: find all prices between and currency
      operationId: getPricesBetweenAndCurrency
      parameters:
        - name: currency
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: path
          required: true
          schema:
            type: number
        - name: to
          in: path
          required: true
          schema:
            type: number
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        200:
          $ref: '#/components/responses/PricesResponse'
        400:
          $ref: '#/components/responses/RequestError'
        401:
          $ref: '#/components/responses/RequestError'
        403:
          $ref: '#/components/responses/RequestError'
        404:
          $ref: '#/components/responses/RequestError'
        500:
          $ref: '#/components/responses/RequestError'
...
  /v1/price/delete/{id}:
    delete:
      tags:
        - price
      operationId: deletePrice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Ok
        400:
          $ref: '#/components/responses/RequestError'
        401:
          $ref: '#/components/responses/RequestError'
        403:
          $ref: '#/components/responses/RequestError'
        404:
          $ref: '#/components/responses/RequestError'
        500:
          $ref: '#/components/responses/RequestError'
...
components:
...
  requestBodies:
    PriceDto:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PriceDto'
    BetweenPrice:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BetweenPrice'
    ProductDto:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductDto'
...
  responses:
...
    PriceResponse:
      description: Price
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PriceDto'
    PricesResponse:
      description: Prices
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/PriceDto'
...
    RequestError:
      description: Request Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'

  #############################################
  ## Schemas
  #############################################
  schemas:
...
    ProblemDetail:
      type: object
    Pageable:
      type: object
      properties:
        dump:
          type: boolean

    # DTO Business
    ProductDto:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        prices:
          type: array
          items:
            type: integer
            format: int64
        categories:
          type: array
          items:
            type: integer
            format: int64
    PriceDto:
      required:
        - amount
        - currency
        - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        currency:
          maxLength: 3
          minLength: 3
          type: string
        amount:
          type: number
        productId:
          type: integer
          format: int64
...
    BetweenPrice:
      required:
        - currency
        - from
        - to
      type: object
      properties:
        from:
          type: number
        to:
          type: number
        currency:
          maxLength: 3
          minLength: 3
          type: string

    # DTO Auth
...